apps/api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
langchain_ollama
langchain_openai
langchain_chroma
pypdf>=5.0.0
motor>=3.3.2
pymongo>=4.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
docx2txt>=0.8